kodilla-spring-basic/src/main/java/com/kodilla/spring/basic_dependency_injection/homework/ShippingCenter.java
package com.kodilla.spring.basic_dependency_injection.homework;

public class ShippingCenter {

    private final DeliveryService deliveryService;
    private final NotificationService notificationService;

    public ShippingCenter(DeliveryService deliveryService, NotificationService notificationService) {
        this.deliveryService = deliveryService;
        this.notificationService = notificationService;
    }

    public String sendPackage(String address, double weight) {
        if (deliveryService.deliverPackage(address, weight)) {
            return notificationService.success(address);
        }
        return notificationService.fail(address);
    }
}